refactor(checker): split @typedef base-type validation out of jsdoc/diagnostics.rs
Move the JSDoc @typedef base-type validation cluster
(check_jsdoc_typedef_base_types and its helpers) into a new sibling
module jsdoc/base_types.rs. Pure code move: the eight impl CheckerState
methods are relocated verbatim. Three module-private helpers that the
moved cluster still calls (jsdoc_file_is_global_script,
jsdoc_param_return_type_spans, jsdoc_bare_param_return_type_spans) are
widened from fn to pub(crate) so the cross-module calls resolve;
visibility widening cannot change runtime behavior.

diagnostics.rs drops 2440 -> 1450 LOC, back under the 2000-LOC checker
boundary; base_types.rs is 1002 LOC. Removes diagnostics.rs from the
arch_guard LINE_LIMIT_CHECKS allowlist (checker over-limit set 6 -> 5)
and tightens its FILE_LINE_LIMIT_CHECKS ratchet to 1450.

AgentName: M1-Opus
